Bug fix for MPU generation & addition of MPU Box dump
